+++ Core/OfficeDevPnP.Core/Framework/Provisioning/ObjectHandlers/ObjectSiteSecurity.cs
@@ -71,7 +71,7 @@ namespace OfficeDevPnP.Core.Framework.Provisioning.ObjectHandlers
 
                         if (setAssociatedOwnerGroup)
                         {
-                            if (parsedAssociatedOwnerGroupName == string.Empty)
+                            if (string.IsNullOrEmpty(parsedAssociatedOwnerGroupName))
                             {
                                 // does throw exception "Value cannot be null" - todo: how to clear the group?
                                 //web.AssociatedOwnerGroup = null;
@@ -114,7 +114,7 @@ namespace OfficeDevPnP.Core.Framework.Provisioning.ObjectHandlers
 
                         if (setAssociatedMemberGroup)
                         {
-                            if (parsedAssociatedMemberGroupName == string.Empty)
+                            if (string.IsNullOrEmpty(parsedAssociatedMemberGroupName))
                             {
                                 // does throw exception "Value cannot be null" - todo: how to clear the group?
                                 //web.AssociatedMemberGroup = null;
@@ -157,7 +157,7 @@ namespace OfficeDevPnP.Core.Framework.Provisioning.ObjectHandlers
 
                         if (setAssociatedVisitorGroup)
                         {
-                            if (parsedAssociatedVisitorGroupName == string.Empty)
+                            if (string.IsNullOrEmpty(parsedAssociatedVisitorGroupName))
                             {
                                 // does throw exception "Value cannot be null" - todo: how to clear the group?
                                 //web.AssociatedVisitorGroup = null;