+++ Core/OfficeDevPnP.Core/Framework/Provisioning/ObjectHandlers/ObjectWebSettings.cs
@@ -64,6 +64,7 @@ namespace OfficeDevPnP.Core.Framework.Provisioning.ObjectHandlers
 #if !SP2019
                 webSettings.SearchScope = (SearchScopes)Enum.Parse(typeof(SearchScopes), web.SearchScope.ToString(), true);
                 webSettings.SearchBoxInNavBar = (SearchBoxInNavBar)Enum.Parse(typeof(SearchBoxInNavBar), web.SearchBoxInNavBar.ToString(), true);
+                webSettings.SearchCenterUrl = web.GetWebSearchCenterUrl(true);
     #endif
 #endif
                 // We're not extracting Title and Description
@@ -391,6 +392,12 @@ namespace OfficeDevPnP.Core.Framework.Provisioning.ObjectHandlers
                     if(web.SearchBoxInNavBar.ToString() != webSettings.SearchBoxInNavBar.ToString())
                     {
                         web.SearchBoxInNavBar = (SearchBoxInNavBarType)Enum.Parse(typeof(SearchBoxInNavBarType), webSettings.SearchBoxInNavBar.ToString(), true);
+                    }
+
+                    if (!string.IsNullOrEmpty(webSettings.SearchCenterUrl) &&
+                        web.GetWebSearchCenterUrl(true) != webSettings.SearchCenterUrl)
+                    {
+                        web.SetWebSearchCenterUrl(webSettings.SearchCenterUrl);
                     }
 #endif
 #endif