+++ Core/OfficeDevPnP.Core/Sites/SiteCollection.cs
@@ -1,4 +1,5 @@
 ﻿#if !ONPREMISES
+using Microsoft.Graph;
 using Microsoft.SharePoint.Client;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
@@ -158,6 +159,10 @@ namespace OfficeDevPnP.Core.Sites
                     optionalParams.Add("Description", siteCollectionCreationInformation.Description ?? "");
                     optionalParams.Add("Classification", siteCollectionCreationInformation.Classification ?? "");
                     var creationOptionsValues = new List<string>();
+                    if (siteCollectionCreationInformation.SiteDesignId.HasValue)
+                    {
+                        creationOptionsValues.Add($"implicit_formula_292aa8a00786498a87a5ca52d9f4214a_{siteCollectionCreationInformation.SiteDesignId.Value.ToString("D").ToLower()}");
+                    }
                     if (siteCollectionCreationInformation.Lcid != 0)
                     {
                         creationOptionsValues.Add($"SPSiteLanguage:{siteCollectionCreationInformation.Lcid}");
@@ -1052,6 +1057,48 @@ namespace OfficeDevPnP.Core.Sites
                 return await Task.Run(() => responseString);
             }
         }
+
+        /// <summary>
+        /// Turns a team site into a communication site
+        /// </summary>
+        /// <param name="context">ClientContext of the team site to update to a communication site</param>
+        /// <returns></returns>
+        public static async Task EnableCommunicationSite(ClientContext context)
+        {
+            await EnableCommunicationSite(context, Guid.Parse("96c933ac-3698-44c7-9f4a-5fd17d71af9e"));
+        }
+
+        /// <summary>
+        /// Turns a team site into a communication site
+        /// </summary>
+        /// <param name="context">ClientContext of the team site to update to a communication site</param>
+        /// <param name="designPackageId">Design package id to be applied, 96c933ac-3698-44c7-9f4a-5fd17d71af9e (Topic = default), 6142d2a0-63a5-4ba0-aede-d9fefca2c767 (Showcase) or f6cc5403-0d63-442e-96c0-285923709ffc (Blank)</param>
+        /// <returns></returns>
+        public static async Task EnableCommunicationSite(ClientContext context, Guid designPackageId)
+        {
+
+            if (context == null)
+            {
+                throw new ArgumentNullException("context");
+            }
+
+            context.Web.EnsureProperty(p => p.Url);
+
+            if (designPackageId == Guid.Empty)
+            {
+                throw new Exception("Please specify a valid designPackageId");
+            }
+
+            if (designPackageId != Guid.Parse("96c933ac-3698-44c7-9f4a-5fd17d71af9e") &&  // Topic
+                designPackageId != Guid.Parse("6142d2a0-63a5-4ba0-aede-d9fefca2c767") &&  // Showcase
+                designPackageId != Guid.Parse("f6cc5403-0d63-442e-96c0-285923709ffc"))    // Blank
+            {
+                throw new Exception("Invalid designPackageId specified. Use 96c933ac-3698-44c7-9f4a-5fd17d71af9e (Topic = default), 6142d2a0-63a5-4ba0-aede-d9fefca2c767 (Showcase) or f6cc5403-0d63-442e-96c0-285923709ffc (Blank)");
+            }
+
+            await context.Web.ExecutePost("/_api/sitepages/communicationsite/enable", $@" {{ ""designPackageId"": ""{designPackageId.ToString()}"" }}");
+        }
+
     }
 }
 #endif